datalogging

+++ app/samples/datalogger/src/main.cpp
@@ -1,16 +1,16 @@
-#include "f_core/os/datalogger.h"
+#include "f_core/os/c_datalogger.h"
 // #include <vecto/r>
 #include <zephyr/kernel.h>
 
 struct PacketA {
-    int a;
-    int b;
+    uint8_t a;
+    uint8_t b;
 };
 CDataLogger<PacketA> alogger{"/lfs/a.bin"};
 
 int main() {
     printk("asdsadsaads\n");
-    for (int i = 0; i < 100; i++) {
+    for (uint8_t i = 0; i < 100; i++) {
         alogger.write({i, 100 - i});
         k_msleep(10);
     }
